Refactor and replace `AssetHandleManager` with `DescriptorManager` for improved resource and descriptor handling
- Removed `AssetHandleManager` and associated resource management logic.
- Introduced `DescriptorManager` to handle Vulkan descriptor pools, layouts, and resources.
- Reorganized asset and Vulkan resource management structure, including the addition of `IVulkanDescriptor`.
- Updated references to `MeshHandle`, `ShaderHandle`, and `NodeHandle` with new descriptor-based classes.
- Adjusted project structure and included necessary updates to CMake and `.idea` configuration.

+++ src/assetManager/src/assets/textureAsset/TextureAsset.h
@@ -32,7 +32,7 @@ namespace am {
         [[nodiscard]] int getChannels() const { return data.channels ; }
         [[nodiscard]] bool hasAlpha() const { return data.hasAlpha; }
 
-        [[nodiscard]] const uint8_t *getData() const {
+        [[nodiscard]] const unsigned* getData() const {
             return data.pixels.data();
         }
